refactor: Extracted functions (update-test-app-package-json)
