+++ src/migration/ember-addon/steps/update-test-app-package-json.ts
@@ -1,86 +1,15 @@
 import { writeFileSync } from 'node:fs';
 import { join } from 'node:path';
 
+import { readPackageJson } from '@codemod-utils/json';
+
+import type { Options } from '../../../types/index.js';
 import {
-  convertToMap,
-  convertToObject,
-  readPackageJson,
-} from '@codemod-utils/json';
-
-import type { Options, PackageJson } from '../../../types/index.js';
-import { getVersion } from '../../../utils/blueprints.js';
-
-function moveDependenciesToDevDependencies(
-  packageJson: PackageJson,
-  options: Options,
-): void {
-  const { packages } = options;
-
-  const dependencies = convertToMap(packageJson['dependencies']);
-  const devDependencies = convertToMap(packageJson['devDependencies']);
-
-  const packagesToMove = new Set([
-    'ember-auto-import',
-    'ember-cli-babel',
-    'ember-cli-htmlbars',
-  ]);
-
-  if (packages.addon.hasTypeScript) {
-    packagesToMove.add('ember-cli-typescript');
-  }
-
-  Array.from(packagesToMove)
-    .filter((packageName) => dependencies.has(packageName))
-    .forEach((packageName) => {
-      const version = getVersion(packageName, options);
-
-      devDependencies.set(packageName, version);
-
-      dependencies.delete(packageName);
-    });
-
-  packageJson['dependencies'] = convertToObject(dependencies);
-  packageJson['devDependencies'] = convertToObject(devDependencies);
-}
-
-function updateDependencies(packageJson: PackageJson): void {
-  const dependencies = convertToMap(packageJson['dependencies']);
-
-  /*
-    For the time being, we'll take the approach of starting over and
-    adding back the dependencies that are required.
-  */
-  dependencies.clear();
-
-  packageJson['dependencies'] = convertToObject(dependencies);
-}
-
-function updateDevDependencies(
-  packageJson: PackageJson,
-  options: Options,
-): void {
-  const { packages } = options;
-
-  const devDependencies = convertToMap(packageJson['devDependencies']);
-
-  devDependencies.set(packages.addon.name, packages.addon.version);
-
-  packageJson['devDependencies'] = convertToObject(devDependencies);
-}
-
-function updateOtherFields(packageJson: PackageJson, options: Options): void {
-  const { packages } = options;
-
-  delete packageJson['ember-addon'];
-
-  packageJson['keywords'] = (packageJson['keywords'] ?? []).filter(
-    (keyword: string) => keyword !== 'ember-addon',
-  );
-
-  packageJson['name'] = packages.testApp.name;
-
-  packageJson['private'] = true;
-}
+  moveDependenciesToDevDependencies,
+  updateDependencies,
+  updateDevDependencies,
+  updateOtherFields,
+} from './update-test-app-package-json/index.js';
 
 export function updateTestAppPackageJson(options: Options): void {
   const { locations, projectRoot } = options;
