+++ src/migration/ember-addon/steps/update-test-app-package-json/index.ts
@@ -1,0 +1,4 @@
+export * from './move-dev-dependencies-to-dependencies.js';
+export * from './update-dependencies.js';
+export * from './update-dev-dependencies.js';
+export * from './update-other-fields.js';
