+++ src/migration/ember-addon/steps/update-test-app-package-json/update-dev-dependencies.ts
@@ -1,0 +1,16 @@
+import { convertToMap, convertToObject } from '@codemod-utils/json';
+
+import type { Options, PackageJson } from '../../../../types/index.js';
+
+export function updateDevDependencies(
+  packageJson: PackageJson,
+  options: Options,
+): void {
+  const { packages } = options;
+
+  const devDependencies = convertToMap(packageJson['devDependencies']);
+
+  devDependencies.set(packages.addon.name, packages.addon.version);
+
+  packageJson['devDependencies'] = convertToObject(devDependencies);
+}
